requirements.txt
ultralytics>=8.3.0
opencv-python>=4.10.0
dxcam>=0.0.5
numpy>=1.26.0
keyboard>=0.13.5
openai>=1.40.0
google-generativeai>=0.8.0
pyyaml>=6.0
pydantic>=2.7.0
python-dotenv>=1.0.1
loguru>=0.7.2
win10toast>=0.9 ; platform_system == "Windows"
